+++ EsercizioUno_Tre.java
@@ -30,7 +30,7 @@ public class EsercizioUno_Tre {
 		    state = 1;
 		else if (ch == '1' || ch == '3' || ch == '5' || ch == '7' || ch == '9')
 		    state = 2;
-		else if ((int)ch >= 65 && (int)ch <= 75)
+		else if (ch >= 65 && ch <= 75)
 			state = 3;
 		else
 		    state = -1;
@@ -41,14 +41,14 @@ public class EsercizioUno_Tre {
 		    state = 2;
 		else if (ch == '0' || ch == '2' || ch == '4' || ch == '6' || ch == '8')
 		    state = 1;
-		else if ((int)ch >= 76 && (int)ch <= 90)
+		else if (ch >= 76 && ch <= 90)
 			state = 3;
 		else
 		    state = -1;
 		break;
 
 	    case 3:
-		if ((int)ch >= 97 && (int)ch <= 122)
+		if (ch >= 97 && ch <= 122)
 		    state = 3;
 		else
 		    state = -1;
